1. 修正帳號資訊，董事長頭銜與公司名稱的顯示；同樣修正公司詳細資訊，買賣單的使用者名稱顯示，改用max-width 2. 修正公司詳細資訊，K線圖下降的顯示顏色 3. 依照一個publication一個檔案原則，將accountCompanyTitle拆成accountChairmanTitle、accountManagerTitle、accountEmployeeTitle
